\name{FOUND}
\alias{FOUND}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Found population
}
\description{
Found a new population of speakers to start a simulation.
}
\usage{
FOUND(nAgents = world$nAgents)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{nAgents}{
number of agents to start with.
}
}
\value{
starting population
}
\author{
Sander Lestrade
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{MULTIRUN}
}
\examples{
FOUND(4)
names(population)
}
\keyword{misc}
